requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langgraph
langchain-openai
psycopg2-binary
